WIP: split map utilities control 293 (checkpoint)
